backend/pyproject.toml
[project]
name = "app-hostify-backend"
version = "0.1.0"
description = "Self-hosted deployment platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "docker>=7.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "websockets>=14.1",
    "aiofiles>=24.1.0",
    "pyyaml>=6.0.2",
    "gitpython>=3.1.43",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


